[IMP] Added workspace-from-pr, renamed commands for standardization, cleanup
`workspace-from-pr` creates a new Workspace from a given PR number.
It accepts the first repository that comes up with a valid PR for the
moment. Starts from there, then looks up the other repos for the
same branch name. Asks for the db_name and the modules, then
creates the Workspace automatically, with the name of the branch.

Workspace methods have now a 'workspace_' prefix instead of a suffix.
Now this has been made coherent in all methods that deal with workspaces.